requirements.txt
django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.8.6
spacy==3.1.1
spacy-legacy==3.0.8
django-cors-headers==3.8.0